fix: use write_all() to write space after a tag
`write()` should not return Ok without writing anything,
but the documentation is not clear on this
and underlying implementation may be wrong
as IMAP can be run over various streams,
e.g. compressed, encrypted, proxied.
It is safer to use `write_all()` everywhere,
even when we write a single byte.