refactor!: use try_next() more
I have looked through the code
to make sure errors are not ignored,
especially in loops.
Ignoring errors in loops
when reading from streams
may result in infinite loop
reading the same error over and over again.
No bugs found,
but I refactored reading from streams
to use `try_next()` more
to bubble up the errors with `?`
as soon as possible.

This is a breaking change since
`read_response()` is resultified to return `Result<Option<_>>`
instead of `Option<Result<_>>`.
`read_response()` is a public interface
that is used by library users
to read the banner.