feat: close ImapStream after an error
This prevents calls to poll_next()
from reaching the underlying stream
after an error is returned once.

Previously calls to poll_next()
to ImapStream built on top
of a network connection
could have returned infinite
stream of Some(Err(_))
values.
This is dangerous
for the code that looks for the end
of stream without processing
errors such as
`while stream.next().await.is_some() {}`
because it may result in
infinite loop.

It is still better to process
errors by writing
`while stream.try_next().await?.is_some() {}`,
but the change protects
in case of incorrect library user code.